sweater

+++ src/main/java/com/example/sweater/repos/MessageRepo.java
@@ -1,9 +1,8 @@
 package com.example.sweater.repos;
-
 import com.example.sweater.domain.Message;
 import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 
-public interface MessageRepo extends CrudRepository<Message, Integer> {
+public interface MessageRepo extends CrudRepository<Message, Long> {
     List<Message> findByTag(String tag);
 }
